Move tests from utils to correct class
